* subversion/libsvn_wc/upgrade.c   (remove_suffix): Use correct type to avoid bit-wise mismatches.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1058632 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_wc/upgrade.c
@@ -942,8 +942,8 @@ migrate_props(const char *dir_abspath,
 static char *
 remove_suffix(const char *str, const char *suffix, apr_pool_t *result_pool)
 {
-  int str_len = strlen(str);
-  int suffix_len = strlen(suffix);
+  size_t str_len = strlen(str);
+  size_t suffix_len = strlen(suffix);
 
   if (str_len > suffix_len
       && strcmp(str + str_len - suffix_len, suffix) == 0)
